Added better game loop and cleaned up
