+++ src/Menus/MainMenu.cs
@@ -1,4 +1,5 @@
-﻿using Spectre.Console;
+﻿using DustToDust.Chapters;
+using Spectre.Console;
 
 namespace DustToDust.Menus;
 
@@ -6,27 +7,33 @@ public static class MainMenu
 {
     public static void WelcomeMessage()
     {
-        Console.Clear();
-        AnsiConsole.Write(Globals.Title);
-
-        var choice = AnsiConsole.Prompt(
-            new SelectionPrompt<string>()
-                .Title("  What do you want to do?")
-                .HighlightStyle(Style.Plain)
-                .AddChoices("Log In", "About", "Exit"));
-
-        switch (choice)
+        var isLogIn = false;
+        var isRunning = true;
+        do
         {
-            case "Log In":
-                LogIn();
-                break;
-            case "About":
-                About();
-                break;
-            case "Exit":
-                Exit();
-                break;
-        }
+            Console.Clear();
+            AnsiConsole.Write(Globals.Title);
+            var choice = AnsiConsole.Prompt(
+                new SelectionPrompt<string>()
+                    .Title("  What do you want to do?")
+                    .HighlightStyle(Style.Plain)
+                    .AddChoices("Log In", "About", "Exit"));
+            switch (choice)
+            {
+                case "Log In":
+                    isLogIn = true;
+                    isRunning = false;
+                    break;
+                case "About":
+                    About();
+                    break;
+                case "Exit":
+                    isRunning = false;
+                    break;
+            }           
+        } while (isRunning);
+        if (isLogIn) LogInChapter.LogIn();
+        else Exit();
     }
 
     private static void About()
@@ -44,7 +51,6 @@ public static class MainMenu
             .Centered();
         AnsiConsole.Write(text);
         Console.ReadKey(true);
-        WelcomeMessage();
     }
 
     private static void Exit()
@@ -55,66 +61,5 @@ public static class MainMenu
             Centered(); // TODO: Custom messages based on current story
         AnsiConsole.Write(text);
         Console.ReadKey(true);
-        Console.Clear();
-        Environment.Exit(0);
-    }
-
-    private static void LogIn()
-    {
-        Console.Clear();
-        Thread.Sleep(1000);
-        BottomLeftAnimated("15:40:06 UTC\nArriving at Jupiter AO");
-        Hear("*cockpit door knocks repeatedly*");
-        Talk(Prompt(["What?", "No.", "*ignore*"]));
-        Console.ReadKey(true);
-    }
-
-    public static void BottomLeftAnimated(string prompt)
-    {
-        Console.Clear();
-        var windowWidth = Console.WindowWidth;
-        var windowHeight = Console.WindowHeight;
-        Console.SetCursorPosition(0, windowHeight - 5);
-        foreach (var c in prompt)
-        {
-            AnsiConsole.Markup($"[italic bold]{c}[/]");
-            Thread.Sleep(25);
-        }
-        Thread.Sleep(1000);
-        Console.Clear();
-        Console.SetCursorPosition(0, 0);
-    }
-
-    public static void Hear(string dialogue)
-    {
-        Console.Clear();
-        Globals.SetHearCursor();
-        foreach (var c in dialogue)
-        {
-            AnsiConsole.Markup($"[italic]{c}[/]");
-            Thread.Sleep(20);
-        }
-        Thread.Sleep(500);
-    }
-
-    public static void Talk(string dialogue)
-    {
-        Globals.SetTalkCursor();
-        AnsiConsole.Markup(">> ");
-        foreach (var c in dialogue)
-        {
-            AnsiConsole.Markup($"{c}");
-            Thread.Sleep(20);
-        }
-    }
-
-    public static string Prompt(string[] options)
-    {
-        Globals.SetPromptCursor();
-        var choice = AnsiConsole.Prompt(
-            new SelectionPrompt<string>()
-                .HighlightStyle(Style.Plain)
-                .AddChoices(options));
-        return choice;
     }
 }
